+++ cubi_sak/__main__.py
@@ -15,6 +15,8 @@ from .isa_tpl import run as run_isa_tpl
 from .isa_tpl import setup_argparse as setup_argparse_isa_tpl
 from .snappy import run as run_snappy
 from .snappy import setup_argparse as setup_argparse_snappy
+from .sodar import run as run_sodar
+from .sodar import setup_argparse as setup_argparse_sodar
 
 
 def setup_argparse_only():  # pragma: nocover
@@ -46,6 +48,7 @@ def setup_argparse():
     setup_argparse_snappy(
         subparsers.add_parser("snappy", help="Tools for supporting the SNAPPY pipeline.")
     )
+    setup_argparse_sodar(subparsers.add_parser("sodar", help="SODAR command line interface."))
 
     return parser, subparsers
 
@@ -70,7 +73,13 @@ def main(argv=None):
     logzero.loglevel(level=level)
 
     # Handle the actual command line.
-    cmds = {None: run_nocmd, "isa-tpl": run_isa_tpl, "isa-tab": run_isa_tab, "snappy": run_snappy}
+    cmds = {
+        None: run_nocmd,
+        "isa-tpl": run_isa_tpl,
+        "isa-tab": run_isa_tab,
+        "snappy": run_snappy,
+        "sodar": run_sodar,
+    }
 
     res = cmds[args.cmd](args, parser, subparsers.choices[args.cmd] if args.cmd else None)
     if not res: